auth

+++ src/main/kotlin/buyticket/auth/repositories/SessionRepository.kt
@@ -3,7 +3,7 @@ package buyticket.auth.repositories
 import buyticket.auth.models.Session
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface SessionRepository : JpaRepository<Session, Int> {
+interface SessionRepository : JpaRepository<Session, Long> {
 
     fun findByUserId(userId: Int): Session?
 }
